1. Add new folders (domain, data, presentation) 2. Add usecases and POJO model 3. Change place of UI layers
