+++ app/src/main/java/com/example/mytestapplication/domain/AddShopItem.kt
@@ -1,0 +1,7 @@
+package com.example.mytestapplication.domain
+
+class AddShopItem {
+    fun addShopItem(shopItem: ShopItem){
+        TODO()
+    }
+}
