+++ app/src/main/java/com/example/mytestapplication/domain/DeleteShopItem.kt
@@ -1,0 +1,7 @@
+package com.example.mytestapplication.domain
+
+class DeleteShopItem {
+    fun deleteShopItem(shopItem: ShopItem){
+        TODO()
+    }
+}
