+++ app/src/main/java/com/example/mytestapplication/domain/EditShopItem.kt
@@ -1,0 +1,10 @@
+package com.example.mytestapplication.domain
+
+class EditShopItem {
+
+    fun editShopItem(shopItem: Int) :  ShopItem{
+        //go to the next page using id (not the whole object) when try to edit shop item
+        TODO()
+    }
+
+}
