+++ app/src/main/java/com/example/mytestapplication/domain/GetShopItemUseCase.kt
@@ -1,0 +1,9 @@
+package com.example.mytestapplication.domain
+
+class GetShopItemUseCase {
+
+    fun  getShopList() : List<ShopItem>{
+        TODO()
+    }
+
+}
