+++ app/src/main/java/com/example/mytestapplication/domain/ShopItem.kt
@@ -1,0 +1,9 @@
+package com.example.mytestapplication.domain
+
+data class ShopItem(
+    val id: Int,
+    val name: String,
+    val description: String,
+    val price: Int
+) {
+}
