+++ app/src/main/java/com/example/mytestapplication/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.mytestapplication
+package com.example.mytestapplication.presentation
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,6 +7,7 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import com.example.mytestapplication.R
 import com.example.mytestapplication.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
